Small refactoring on StructuredProgram, move RemovePhis method to a separate class
